Fix requests to URLs like proposal/add/?call=18as
call parameter was expected to be a number. Now if it starts as a number
it works. If it doesn't start as a number (it's a purely string) it's
handled.

There is some link on the internet mal-functioning causing other alerts
in the application.

(cherry picked from commit 833a16cb293ada4539706d93b1eb463e66216c59)